- US-3.5 implemented, reviewed, tested. - f64 to Decimal for all money handling

+++ worker/src/client.rs
@@ -1,5 +1,6 @@
 use anyhow::{Context, Result};
 use reqwest::{Client as HttpClient, StatusCode};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::sync::Arc;
@@ -180,7 +181,7 @@ impl WorkerApiClient {
         activity_id: Uuid,
         worker_id: &str,
         output: Value,
-        cost_usd: Option<f64>,
+        cost_usd: Option<Decimal>,
     ) -> Result<()> {
         let token = self.get_token().await?;
 
@@ -189,7 +190,7 @@ impl WorkerApiClient {
             worker_id: String,
             output: Value,
             #[serde(skip_serializing_if = "Option::is_none")]
-            cost_usd: Option<f64>,
+            cost_usd: Option<Decimal>,
         }
 
         let response = self
